requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0

# Machine Learning - Lighter alternatives
# We'll comment out the heavier packages for now
# tensorflow==2.12.0
# keras==2.12.0
# torch==2.0.1

# MLOps Tools
mlflow>=2.0.0
# dvc>=2.0.0
# prefect>=2.0.0
# airflow>=2.0.0

# Data Visualization
matplotlib>=3.0.0
# seaborn>=0.12.0
# plotly>=5.0.0

# Model Serving
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
PyYAML>=6.0.0
requests>=2.28.0

# Development
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
# jupyter>=1.0.0

# Containerization
# docker>=6.0.0

# Kubernetes
kubernetes>=26.0.0

# Prometheus Client
prometheus-client==0.16.0 